TASK: Move `PropertyNames` next to `PropertyName`
